refactor(api): [#208] use API impementation enum for API versioning
We've removed the ActixWeb implementation for the API. We can use the
enum for implementations for versioning the API. Currently there is only
one version `v1` but i'ts ready to add the verion `v2`.
