+++ src/bin/main.rs
@@ -1,16 +1,16 @@
 use torrust_index_backend::app;
 use torrust_index_backend::bootstrap::config::init_configuration;
-use torrust_index_backend::web::api::Implementation;
+use torrust_index_backend::web::api::Version;
 
 #[tokio::main]
 async fn main() -> Result<(), std::io::Error> {
     let configuration = init_configuration().await;
 
-    let api_implementation = Implementation::Axum;
+    let api_version = Version::V1;
 
-    let app = app::run(configuration, &api_implementation).await;
+    let app = app::run(configuration, &api_version).await;
 
-    match api_implementation {
-        Implementation::Axum => app.axum_api_server.unwrap().await.expect("the Axum API server was dropped"),
+    match api_version {
+        Version::V1 => app.api_server.unwrap().await.expect("the API server was dropped"),
     }
 }
