+++ src/web/api/mod.rs
@@ -3,7 +3,7 @@
 //! Currently, the API has only one version: `v1`.
 //!
 //! Refer to the [`v1`](crate::web::api::v1) module for more information.
-pub mod axum;
+pub mod server;
 pub mod v1;
 
 use std::net::SocketAddr;
@@ -14,20 +14,17 @@ use tokio::task::JoinHandle;
 use crate::common::AppData;
 use crate::web::api;
 
-pub const API_VERSION: &str = "v1";
-
-/// API implementations.
-pub enum Implementation {
-    /// API implementation with Axum.
-    Axum,
+/// API versions.
+pub enum Version {
+    V1,
 }
 
 /// The running API server.
 pub struct Running {
     /// The socket address the API server is listening on.
     pub socket_addr: SocketAddr,
-    /// The handle for the running API server task when using Axum.
-    pub axum_api_server: Option<JoinHandle<Result<(), std::io::Error>>>,
+    /// The handle for the running API server.
+    pub api_server: Option<JoinHandle<Result<(), std::io::Error>>>,
 }
 
 #[must_use]
@@ -38,8 +35,8 @@ pub struct ServerStartedMessage {
 
 /// Starts the API server.
 #[must_use]
-pub async fn start(app_data: Arc<AppData>, net_ip: &str, net_port: u16, implementation: &Implementation) -> api::Running {
+pub async fn start(app_data: Arc<AppData>, net_ip: &str, net_port: u16, implementation: &Version) -> api::Running {
     match implementation {
-        Implementation::Axum => axum::start(app_data, net_ip, net_port).await,
+        Version::V1 => server::start(app_data, net_ip, net_port).await,
     }
 }
