--- tests/e2e/contexts/proxy/contract.rs
@@ -1,6 +1,0 @@
-//! API contract for `proxy` context.
-
-mod with_axum_implementation {
-
-    // todo
-}
