+++ tests/e2e/environment.rs
@@ -1,6 +1,6 @@
 use std::env;
 
-use torrust_index_backend::web::api::Implementation;
+use torrust_index_backend::web::api::Version;
 
 use super::config::{init_shared_env_configuration, ENV_VAR_E2E_SHARED};
 use crate::common::contexts::settings::Settings;
@@ -55,7 +55,7 @@ impl TestEnv {
 
     /// It starts the test environment. It can be a shared or isolated test
     /// environment depending on the value of the `ENV_VAR_E2E_SHARED` env var.
-    pub async fn start(&mut self, api_implementation: Implementation) {
+    pub async fn start(&mut self, api_version: Version) {
         let e2e_shared = ENV_VAR_E2E_SHARED; // bool
 
         if let Ok(_e2e_test_env_is_shared) = env::var(e2e_shared) {
@@ -66,7 +66,7 @@ impl TestEnv {
             self.starting_settings = self.server_settings_for_shared_env().await;
         } else {
             // Using an isolated test env.
-            let isolated_env = isolated::TestEnv::running(api_implementation).await;
+            let isolated_env = isolated::TestEnv::running(api_version).await;
 
             self.isolated = Some(isolated_env);
             self.starting_settings = self.server_settings_for_isolated_env();
