+++ tests/e2e/web/api/mod.rs
@@ -1,0 +1,1 @@
+pub mod v1;
