+++ tests/e2e/web/api/v1/mod.rs
@@ -1,0 +1,1 @@
+pub mod contexts;
