+++ tests/e2e/web/mod.rs
@@ -1,0 +1,1 @@
+pub mod api;
